♻️ Extracted sync RPC calls to separate file
