+++ client-rpc/src/rpc.rs
@@ -1,0 +1,1 @@
+pub mod sync_rpc;
