+++ client-rpc/src/server.rs
@@ -1,8 +1,9 @@
-use std::net::SocketAddr;
-
 use failure::ResultExt;
 use jsonrpc_core::{self, IoHandler};
 use jsonrpc_http_server::{AccessControlAllowOrigin, DomainsValidation, ServerBuilder};
+use secstr::SecUtf8;
+use serde::{Deserialize, Serialize};
+use std::net::SocketAddr;
 
 use chain_core::tx::fee::LinearFee;
 use client_common::error::{Error, ErrorKind, Result};
@@ -18,6 +19,7 @@ use client_index::synchronizer::ManualSynchronizer;
 use client_network::network_ops::DefaultNetworkOpsClient;
 
 use crate::client_rpc::{ClientRpc, ClientRpcImpl};
+use crate::rpc::sync_rpc::{SyncRpc, SyncRpcImpl};
 use crate::Options;
 
 type AppSigner = DefaultSigner<SledStorage>;
@@ -97,12 +99,15 @@ impl Server {
     }
 
     pub fn start_client(&self, io: &mut IoHandler, storage: SledStorage) -> Result<()> {
-        let wallet_client = self.make_wallet_client(storage.clone());
+        let client_rpc_wallet_client = self.make_wallet_client(storage.clone());
+        let sync_rpc_wallet_client = self.make_wallet_client(storage.clone());
         let ops_client = self.make_ops_client(storage.clone());
         let synchronizer = self.make_synchronizer(storage.clone());
         let client_rpc =
-            ClientRpcImpl::new(wallet_client, ops_client, synchronizer, self.network_id);
+            ClientRpcImpl::new(client_rpc_wallet_client, ops_client, self.network_id);
+        let sync_rpc = SyncRpcImpl::new(sync_rpc_wallet_client, synchronizer);
         io.extend_with(client_rpc.to_delegate());
+        io.extend_with(sync_rpc.to_delegate());
         Ok(())
     }
 
@@ -140,4 +145,10 @@ pub(crate) fn rpc_error_from_string(error: String) -> jsonrpc_core::Error {
         message: error,
         data: None,
     }
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct WalletRequest {
+    pub name: String,
+    pub passphrase: SecUtf8,
 }
